Created generic API endpoint for updating instance settings
This makes sense on the API end, because the API should reflect the
underlying data resources fairly closely. I've chosen the word
"organization" instead of "instance" for the API because it's user
facing, and while "instance" is a word that makes sense from our
perspective, but "organization" or "group" or something makes more sense
from a user perspective in terms of how they visualize themselves and
their instance. I'm open to changing this word or approach if there are
compelling reasons though.