Feature/edit website config info (#26)
* Small cleanup

* New file upload component (with some tests)

* Created domain name validation schema

* Custom domain set to proper validation

* Moved isBigger to a separate utils file and wrote some tests

* Tests for file size util function

* Website config editing form

* Created generic API endpoint for updating instance settings

This makes sense on the API end, because the API should reflect the
underlying data resources fairly closely. I've chosen the word
"organization" instead of "instance" for the API because it's user
facing, and while "instance" is a word that makes sense from our
perspective, but "organization" or "group" or something makes more sense
from a user perspective in terms of how they visualize themselves and
their instance. I'm open to changing this word or approach if there are
compelling reasons though.

* Added config editing form and handler and server side logic

* Patch MediaQuery from svelte/reactivity instead of runed

This is a breaking change in the runed library. I thought it had been
fixed in a patch elsewhere in main, but it seems like either it hadn't,
or this branch predates that. Either way I've included it here and will
resolve any conflicts that exist

* Improvements to image upload widget widget

* Moved some hard coded things to props
* Added displaying the image with a reset button
* Made the value property bindable
* Fixed a bug where the ID was set on the wrong element

* Using uuidv4 instead of crypto in client-side code

* Removed custom domain and converted logos to image upload widgets