Feature/inland i18n (#33)
* i18n check on github actions

* Added inlang's sherlock to recommended extensions

* Some first test messages

* Inlang-generated project folder

* Adding inlang to main hook

* Inlang plugin to vite

* Added inlang libraries

* Last files with initial test translations

* Added all language support to settings

* Added inlang NPM run scripts

* Properly import language types in global types

* Updated the LanguageSelect component and form handler to also trigger Paraglide select

The existing LanguageSelect comonent (and form action handler) triggered
a change in the existing handmade i18n system. This one keeps that logic
in place (so the existing i18n system will continue to work) but makes
also triggers the change in the Paraglide system.

* Improved typing on base i18n functions

* Improved typing of language functions in main request handler

* Implement client side handling of paraglide language

* Moved the client side code from +layout.svelte to a separate component

* Upgrading to Paraglide 2

* Deleted paraglide logic and moved to separate component

* Integrated Paraglide server side hooks

The implementation required using AsyncLocalStorage and splitting it out
into a separate hook, which sets  the language then runs the rest of the
handle hook in an AysncLocalStorage context, in order to avoid race
conditions when we're handling multiple requests with different
languages at the same time.

* Updated test runner to latest node version (required for paraglide 2.0)

* Fixing the bug caused by the merge conflict

I had updated the page import to use the new Svelte 5 import rather than
the old $page store in the base +layout.svelte. However, when I resolved
merge conflicts it reintroduced the $page syntax in a couple of places.

* Exclude paraglide codegen from test coverage reporting