Chore/rationalise image uploads (#35)
* Moved all components that used FileUpload component to new SimpleFileUpload component

* Renamed simple_image_upload to simple_file_upload

It can upload more than just images

* Allow setting custom image upload link

Before uploading to S3, we need to call an API function to generate a
signed PUT url. The link we need depends on the bucket we're uploading
to, so this is allowing different to configure which endpoint you call
to generate the link. This brings it in line with the previous file
upload comp.

* Change the UploadWidget to use SimpleFileUpload component

The upload widget manages adding and displaying new website uploads (eg:
petition or event feature images, etc). This is built on top of the
FileUpload component, but with additional logic and UI. This commit
changes it to use the SimpleFileUpload component insrtead.

* Updating to reference renamed SimpleFileUpload component

This was the only component using the previous simple image upload
component that I have renamed to simple file upload.

* Delete the old FileUpload component

* Update comments in SimpleFileUpload to better describe the component