Removed error in message type conversion
We don't actually want to keep the interactive message parameters when
we're converting an interactive message into a type. Before this commit,
this function would convert the message type to "text" but also keep all
the interactive state, too. This doesn't cause an issue in our editing
screens, but would cause 400 errors if sent to the WhatsApp API