Feature/update email templates (#37)
* chore/unrelated: removed unused export due to compiler warning

* unrelated: Added localization object to email generation functions

* Updated event email templates to use handlebars helper functions for event times and locations

* unrelated/fix: Temporary fix to validation errors

This is a tough one. The "no default" timestamp validation rule was
causing blocking 400 errors on page load for /events/new. Changing it to
this helped, but introduced a default event time of today/today's date,
which isn't what we want. Will need to dive into it more carefully, but
this is a fix for the time being. It doesn't appear to be currently
broken in production.

* Removed unsubscribe links from default email templates