Simplify parsing to remove incorrect validation error that was causing bug
Changes in the underlying file upload component created a bug in which
parsing for the uploads would fail silently, and the file wouldn't be
added as a website.upload row in the database, but would still be
uploaded to S3 and previewed.

This fixes it by just parsing the input from the file upload component
directly. Because the new component is typesafe, the parseUnknownData
helper function that was causing the problem is no longer needed.