Updated params of function to queue event emails
The same function is used for reminder emails and follow up emails.
Before this change, the function had no sense of what context it was
being called in, and *only* updated the flag on the event to indicate
that follow up emails had been sent. This resulted in the cron job to
send event notifications triggering multiple times for reminder emails
because the flag indicating the emails had been sent was not set for
reminder emails. Additionally, it *was* incorrectly set for follow up
emails, which meant that follow up emails (had they been scheduled)
would never be sent.

This commit simply updates the function signatures and the invocations.
A different commit will edit the function body to actually enact the
change.