Add new params to event update function
We recently changed the event update function and it didn't get updated
here. This mant that when it was called here, the function would throw a
ReferenceError, which was not handled, which resulted in th queue job
being retried (up to 25 times) which contributed to the high volumes of
mail sent in the recent incident. This change means that the function
should not thorw an error under normal circumstances.