Patch/reminder email incident (#45)
* Updated params of function to queue event emails

The same function is used for reminder emails and follow up emails.
Before this change, the function had no sense of what context it was
being called in, and *only* updated the flag on the event to indicate
that follow up emails had been sent. This resulted in the cron job to
send event notifications triggering multiple times for reminder emails
because the flag indicating the emails had been sent was not set for
reminder emails. Additionally, it *was* incorrectly set for follow up
emails, which meant that follow up emails (had they been scheduled)
would never be sent.

This commit simply updates the function signatures and the invocations.
A different commit will edit the function body to actually enact the
change.

* Add new params to event update function

We recently changed the event update function and it didn't get updated
here. This mant that when it was called here, the function would throw a
ReferenceError, which was not handled, which resulted in th queue job
being retried (up to 25 times) which contributed to the high volumes of
mail sent in the recent incident. This change means that the function
should not thorw an error under normal circumstances.

* Simply the running of SQL query

Remove the extra step to generate the SQL and log it.

* Set the "reminder/followup sent" flag BEFORE sending the emails

This means that if the emails fail, or an error occurs later somewhere
in this handler, the emails won't be sent again. It's better for
reminder *not* to be sent than to be sent too many times.

* Set the correct flag depending on what function was triggered

This change, combined with another commit editing the function
parameters of this function, mean that the correct flags will be set on
the event object depending on which job is being run.

* Trigger the correct queue function depending on the context

Whether a reminder email or a followup email is being sent, choose the
right queue function