Experiment/whatsapp action UI mockup (#40)
* Added shadcn DropdownMenu component

* Added an optional reset callback function for when the upload is deleted

* Adding ellipsis to our string clamp function

* Updating the page to use new WhatsApp thread composition components

* Add some extra AWS addresses to connect-src CSP directive for uploads

* Chore: version bump bits-ui

This happened when installing the latest version of the DropdownMenu
component

* Added default text to NewMessage creation

This is the only part of the new system that uses an old comp. I will
probably refactor it, but for now it's working OK. I just wanted to add
a default message.

* Simple component for displahying message content

It needs to figure out what kind of message it is, and display the
content accordingly

* Component for displaying message buttons

This is only for displaying the buttons, not editing them or anything
like that. Therefore, it just needs to figure out if the message object
has buttons and to display them

* Component that represents a single message

* Compnent to represent a list of messages in a thread

This is the place that owns and  manages the state of the message thread

* Main logic functions for WhatsApp message thread composition

* Editing frame for WhatsApp message content, using shadcn Drawer comp

* Main editing components for editing WhatsApp message content and structure

* Components for managing WhatsApp message actions

* Update the design of the message display components

This small update brings the message display components into line with
WhatsApp messages

* Update config

* Help fix some dev server instability

* Localizations for message creation

Because I ran the autogenerate script this also includes localization
for an unrelated "choose your language" localization

* Removed fetchMessages from messages because it's duplicated in threads.ts

* Functions to work with message components

* Functions to work with message threads

* WhatsApp message template selector

* Removed old component files

* Updated import to refrence new actions files

* Updated import to reference new actions files

* Added toast messages and loading effect on message save

* Version bumped bits-ui

* Component to render a template

* Component to edit the action on a template

Based on the component to edit the action on a message, but underlying
data structure is slightly different. This isn't very DRY, but it may be
better to keep these components separate. We can revisit this if we fund
ourselves having to update them frequently

* The drawer component for editing template content

The UI for editing template tags isn't perfect, because when the drawer
slides out you're not able to see the message content. I think
eventually editing the text components inline will be a simpler
interface, but this is acceptable for now.

* Core component for templates

* New version of component to create message

* New component to select the type of action for a button

* New component for adding actions to a template button

* Making the +page.svelte component work for new system

- Update imports to be new components
- Move most of the logic out to external /actions/*.ts files
- Add loading indicators and better error handling
- Add more consistent toast messages on success and error

* Refactor to avoid mutating function param state

* Removed error in message type conversion

We don't actually want to keep the interactive message parameters when
we're converting an interactive message into a type. Before this commit,
this function would convert the message type to "text" but also keep all
the interactive state, too. This doesn't cause an issue in our editing
screens, but would cause 400 errors if sent to the WhatsApp API

* Tests for WhatsApp message and template logic