Default to western-type name form (#48)
Previously if the country wasn't jp or us, it would not display a name
form at all. This was a holdover from testing, and I only just realised
it was causing problems in production. This patch means that unless the
country is jp, it will display the western order name form.