Uodated bits-ui to @next.98
This release appears to have fixed the underlying bug which is to do
with a race condition in onfocus functions between dialog and dropdown
components leading to an infinite loop in some circumstances.

I feel a bit uneasy because I was never able to create a minimal
reproduction outside the context of the Belcoda app, and the issues
described by other people weren't exactly what we had been experiencing,
however, it was very close to what we experienced (and in the exact same
circumstances, ie: a dropdown inside a dialog).

The bits-ui maintainer says that release 93 finally fixes this
long-standing bug (see:
https://github.com/huntabyte/bits-ui/issues/913). Prior to this version
bump, I was able to experience the bug fairly reliably (but not always).
Since the version bump I've not experienced it once.