Display timezone on event pages
This shows the timezone for the event on the event page. I'm not sure
whether it's necessary for most users, however it was a request for
Roots and it makes sense for them as an international network of
organiations all operating in different timezones.

I have implemented it in a way that will display on all events, but in
the future, it might make sense to only display it on online events or,
alternatively, make it a flag in the event settings.