Feature/logs to sentry (#58)
* Installd pino sentry transport library

* Updated sentry init to use public env var instead of having DNS hardcoded

* [test/unrelated] Updated DB function to use pino logging instead of console.log

This was mostly as a test of the new logging transport, and also because
using console.log() instead of pino for logging this was an oversight.

* Updated pino logger initializer

This has two elements:
1. Use the Pino Sentry log transport in production
2. Move from generating a new instance of Pino for every version of the
logger to using a single instance and using Pino child logger instead.
This will improve performance and remove a potential memory leak with
event emitters being generated by every single server-side module

* Updated all instances of the logger initialization to use import.meta.url

Previously we would write the file name/path identifier manually. I
didn't know about import.meta.url which provides the full absolute file
path automatically. This also means that the file generating the log
will be linked in the development console, and can be opened with
cmd+click, which is a nice bonus.

Going forward, we can just initialize the logger with `import {pino}
from '$lib/server'; const log = pino(import.meta.url);` on any page.