Created some sample test data from the database to use for certain functions
I'm not sure if this is the best approach (of course, we will want
integration tests in the longer term). But for now, it's better than
having sample data pasted into mocks in every separate test file. Feel
free to add to this or to take data from it when writing tests, until we
arrive at a better approach.