Fix schema of interactive messages (and update components and handlers which use them)
Basically, we had the wrong shape of the message object for interactive
messages (ie: messages that are not a template message, but which
contain buttons). As such, our messages were being rejected by the
YCloud API. This is a fix to the schema, and the components for building
interactive messages in the thread builder, in order to create the right
shape, as well as the incoming message handler, to make sure the
action_id is currently parsed from the incoming payload.