Remove "signups" CRUD function
This was a file with functions for `events.signups`, a table which
doesn't exist (signups are tracked in `event.attendees` -- apologies,
the naming might be a bit confusing).

I have moved the only function from this file that had any references,
which was to do with signing up to an event from WhatsApp to the
attendees CRUD functions file, updated the reference in the incoming
webhook handler, and deleted the whole signups file.