Remove queue function to send registration email
This happens anyway in the `create()` function for event attendees. If
send_notifications is set to true (which it is here) then an email
notification will be queued to be sent inside that function. Having this
code here results in the email being sent twice. I think overall having
the code to send the notification live in the same function that creates
the event signup makes sense to avoid having to manually implement it
each time.