Patch/whatsapp events signups branch fixes (#62)
* Fix tests for interactive message types

Changes to the schema of the interactive message object to match the
requirements of the YCloud API meant the tests failed. I've updated the
tests to match thenew schema.

* Remove "signups" CRUD function

This was a file with functions for `events.signups`, a table which
doesn't exist (signups are tracked in `event.attendees` -- apologies,
the naming might be a bit confusing).

I have moved the only function from this file that had any references,
which was to do with signing up to an event from WhatsApp to the
attendees CRUD functions file, updated the reference in the incoming
webhook handler, and deleted the whole signups file.

* Remove queue function to send registration email

This happens anyway in the `create()` function for event attendees. If
send_notifications is set to true (which it is here) then an email
notification will be queued to be sent inside that function. Having this
code here results in the email being sent twice. I think overall having
the code to send the notification live in the same function that creates
the event signup makes sense to avoid having to manually implement it
each time.