Deleting admins (#68)
* Add deleted at to admins table schema

* Add localization messages

* Update admin listing to exclude deleted admins

* Update read statements to ignore deleted entries

* Add function to delete admins

* Add delete button and functionality to delete admins

* Expire admin sessions after they are deleted

* Flash messages after deleting admin

* get correct county when listing

* Add task to reassign admin's resources

* Fix test

* Add mock admin for tests

* Queue task to reassing admin's resources

* add test case

* Removed erroneous localization import

t isn't exported by $lib/server, it needs to be imported from
event.locals as a function params. However, in transitioning to the new
localization system (inlang) we will no longer need the deprecated
localization system

* Stop admin from deleting themselves

It should not be possible to delete the currently signed in admin
account. Making it possible might lead to complications, and potentially
lead to increased support burden (eg: i locked myself out, the only
other admin is an invalid email address, etc)

* Remove SL export from $lib/server that caused circular dependency crashes

Note: This commit also removes the erroneous import of t from
$lib/server

* Update route to include currentlySignedInAdmin

* Update tests to check signedInAdmin

* Expire sessions in session.ts rather than directly in admins.ts

* Change queue job to run as currently signed in admin

Previously it was trying to run as the admin record which was
immediately deleted, which causes it to automatically fail when the
queue handler tries to retrieve the admin record

* Add future note about moving this into a transaction

* Return error if deleting instance.settings.default_admin_id

* Moved error messages to use inlang

* Move cache invalidation after the delete db query

This is to avoid a highly-unlikely race condition edge case where the
cache was deleted, then recreated again by another request before the
database delete was run.

* Update machine translation

---------

Co-authored-by: Django Merope <dmerope@gmail.com>