Change all references to phone_number_id to become phone_number
This is what it was actually referring to, we shouldn't have been
calling it a phone_number_id when it's actually just a phone number.

The reason it was previously phone_number_id, is because that's the
identifier used by Meta's WhatsApp Business Cloud API. YCloud uses phone
numbers directly, so we should update the naming to reflect that.