Refactor the incoming whatsapp message hook handler
This is effectively a rewrite from scratch of this file in order to
achieve the aims of
https://www.notion.so/belcoda/RFC-001-2-Incoming-WhatsApp-messages-1c423474e3b380179735ccdcc0981a9b?pvs=4

As well as implementing our new strategy for deriving instances and
person info from incoming messsages, addition goals of the rewrite were
to make it simpler and much more readable, observable and testable, as
well as to make building in support for different types of messages
easier down the track.

It's a big change in a very important file, so I've tried to add lots of
comments and indications of the direction behind the decisions being
made.