Add some comments for future refactor
These are cases where parsing might seem like overkill because it's
being passed from one internal file to another with no external systems
in between. However, not parsing here removes our type safety, so
changes in the underlying schema won't be picked up by Typescript as
errors here. In the future, we should refactor this to be a union type
in order to cover all cases and still get the type safety we need.