New upsert function for conversations
The WhatsApp conversation window can be updated with timing based on
when conversations begin and are updated. So an upsert function made
sense. However, with WhatsApp moving to a per-message pricing model,
it's unlikely we'll use this in reality. In the future, we might remove
the concept of "conversations" from the app entirely.