Patch/international event signups (#76)
* Allow non-strict country code parsing of phone numbers on event and petition signup via the website

Previously it was strict (ie: if the phone number couldn't be parsed as
a valid phone number in the same country as the instance, it would fail
the signup.

For an international group like Roots, this is a big problem. This is a
good fix for now, but we will have to update the event signup logic
later to truly meet Roots needs (eg: support asking different questions,
collecting additional info, etc)

For now, this is a simple fix which I have tested in development and
appears to work well to allow any people to sign up regardless of phone
number.

* Fix a couple of multiple imports

I guess these must have been mistaken merge conflict issues in a
previous PR

* Move husky to non-dev dependencies

Previously, when in dev depencies it would attempt to trigger in the run
step in Docker when `npm run ci --omit dev` is run and crash the build.
This appears to work.