Improve logging and ensure that the response is always successful
If the response is not always 200, then the queue handler will retry the
send message job. Perhaps one day if we were sure that our code was
bulletproof that might be desired behaviour, but for now it is very
likely to cause more harm than good. And combined with good logging and
Sentry, we should be able to keep track of errors happening this way.