Update sent_email schema to add send details
These details mean we still have a record of exactly which email was
sent and its contents (because they are recreatable from the type of
email send it was and the details contained in the send details). For
example, an event_registration email with a particular event ID tells us
exactly what we need to know: which template was sent and the event it
was sent for. Previously we had this information with a linked
email_message_id, but with system-generated emails no longer having a
corresponding email_message resource, this was required.